+++ src/main/java/org/uqbar/project/wollok/game/gameboard/GameboardListener.java
@@ -1,4 +1,4 @@
-package org.uqbar.project.wollok.game.domain.gameboard;
+package org.uqbar.project.wollok.game.gameboard;
 
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.Gdx;
